# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyTagRequest": {
        "properties": {
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "title": "ApplyTagRequest",
        "type": "object"
      },
      "ApplyTagResult": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "integer"
          },
          "cluster_id": {
            "title": "Cluster Id",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "cluster_id",
          "tag",
          "applied",
          "skipped"
        ],
        "title": "ApplyTagResult",
        "type": "object"
      },
...
      "ClusterDetail": {
        "properties": {
          "cluster_id": {
            "title": "Cluster Id",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "doc_count": {
            "title": "Doc Count",
            "type": "integer"
          },
          "is_noise": {
            "default": false,
            "title": "Is Noise",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "level": {
            "default": 1,
            "title": "Level",
            "type": "integer"
          },
          "parent_cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Cluster Id"
          },
          "parent_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Label"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "top_tags": {
            "items": {
              "type": "string"
            },
            "title": "Top Tags",
            "type": "array"
          }
        },
        "required": [
          "cluster_id",
          "label",
          "description",
          "run_id",
          "doc_count",
          "top_tags"
        ],
        "title": "ClusterDetail",
        "type": "object"
      },
      "ClusterOut": {
        "properties": {
          "cluster_id": {
            "title": "Cluster Id",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "doc_count": {
            "title": "Doc Count",
            "type": "integer"
          },
          "is_noise": {
            "default": false,
            "title": "Is Noise",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "level": {
            "default": 1,
            "title": "Level",
            "type": "integer"
          },
          "parent_cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Cluster Id"
          },
          "parent_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Label"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          }
        },
        "required": [
          "cluster_id",
          "label",
          "description",
          "run_id",
          "doc_count"
        ],
        "title": "ClusterOut",
        "type": "object"
      },
      "ClusterPatchRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "ClusterPatchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local-first research library API",
    "title": "Alexandria",
    "version": "0.0.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clusters/{cluster_id}": {
      "get": {
        "operationId": "get_cluster_clusters__cluster_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cluster",
        "tags": [
          "clusters"
        ]
      },
      "patch": {
        "description": "Persist a manually edited cluster label (and optional description).",
        "operationId": "patch_cluster_clusters__cluster_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Cluster",
        "tags": [
          "clusters"
        ]
      }
    },
    "/clusters/{cluster_id}/apply-tag": {
      "post": {
        "description": "Apply the cluster label (slugified) or an override tag to every document.",
        "operationId": "apply_cluster_tag_clusters__cluster_id__apply_tag_post",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApplyTagRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyTagResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Cluster Tag",
        "tags": [
          "clusters"
        ]
      }
    },
    "/clusters/{cluster_id}/documents": {
      "get": {
        "operationId": "cluster_documents_clusters__cluster_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cluster Documents",
        "tags": [
          "clusters"
        ]
      }
    },
    "/clusters/{cluster_id}/regenerate-label": {
      "post": {
        "description": "Re-run LLM cluster labelling for one cluster.",
        "operationId": "regenerate_cluster_label_clusters__cluster_id__regenerate_label_post",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Cluster Label",
        "tags": [
          "clusters"
        ]
      }
    },
...
  }
}